% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_loess.R
\name{trend_loess}
\alias{trend_loess}
\title{trend_loess: fit loess model and return results of interest}
\usage{
trend_loess(df, x = "cycle", y, ID = "Plot_Name", span = NA_real_)
}
\arguments{
\item{df}{Data frame containing a column called Plot_Name, a column called cycle, and a column with at least one
response variable.}

\item{x}{Quoted time variable for trend analysis. Default is "cycle", but can also model by year. Must be numeric.}

\item{y}{Quoted response variable in the data frame.}

\item{ID}{Quoted name of column containing site or plot IDs. Default is "Plot_Name", and assumes the first 4 characters
are a park code.}

\item{span}{numeric value that controls the degree of smoothing. Smaller values (e.g., 0.1) result in less smoothing,
and possibly over-fitting the curve. Higher values (e.g., 1) result is more smoothing and possibly under-fitting. If no
span is specified, loess.as() will be used to determine the optimum span (Note: user specified is preferred).}
}
\description{
fit a loess model using Plot_Name, and return predicted responses for each time steps represented in the
data. This is mostly an internal function run within case_boot_lmer(), but can be used as a stand alone function.
}
\examples{
\dontrun{
library(magrittr)
fake_df <- data.frame(Plot_Name = rep(paste0(rep("PARK.", 9), 1:9), each = 3),
                      group = "PARK",
                      cycle = rep(1:3, times = 9),
                      resp = runif(27, 0, 20))

fake_df$resp_trend <- fake_df$cycle * fake_df$cycle

mod <- trend_loess(fake_df, x = "cycle", y = "resp", ID = "Plot_Name", span = 1)
}

}
