% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_bootstrap_funs.R
\name{case_boot_lmer}
\alias{case_boot_lmer}
\title{case_boot_lmer: run case bootstrap for random intercept model and return model output}
\usage{
case_boot_lmer(df, y, num_reps, chatty = TRUE)
}
\arguments{
\item{df}{Data frame containing a column called Plot_Name, a column called cycle, and a column with at least one
response variable.}

\item{y}{Quoted response variable in the data frame.}

\item{num_reps}{Number of replicates to run in the bootstrap}

\item{chatty}{TRUE or FALSE. TRUE (default) will print progress in the console, including the first four characters
in the Plot_Name and a tick for every other replicate of the bootstrap. FALSE will not print progress in console.}
}
\description{
For each replicate, as specified by num_reps, function will generate a bootstrapped sample of data from the
original dataset, fit a random intercept model for plot, and return the estimates for intercept, slope, and predicted
values for mean response of each cycle, along with 95 % confidence intervals for each of these terms. The num_boots column
is the number of bootstrapped samples that successfully fit an lmer model. If any singular fits are returned, a warning
message is printed in the console to indicate number of bootstraps that returned singular fits, but the model results are
stored and included in the confidence interval estimates.
}
\examples{
\dontrun{

fake_df <- data.frame(Plot_Name = rep(paste0(rep("PARK.", 9), 1:9), each = 3),
                      cycle = rep(1:3, times = 9),
                      resp = runif(27, 0, 20))

boot1 <- case_boot_lmer(fake_df, y = "resp", num_reps = 10, chatty = TRUE)

}

}
