% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_boot_sample.R
\name{case_boot_sample}
\alias{case_boot_sample}
\title{case_boot_sample: bootstrap resampling function}
\usage{
case_boot_sample(df, y, sample = TRUE, sample_num = 1)
}
\arguments{
\item{df}{Data frame containing a column called Plot_Name, a column called cycle, and a column with at least one
response variable.}

\item{y}{Quoted response variable in the data frame.}

\item{sample}{TRUE or FALSE. TRUE (default) will generate a bootstrapped sample of the specified data frame. FALSE will
run trend_fun() on the original dataset.}

\item{sample_num}{Used for iteration to indicate the replicate number of the bootstrap. Do not need to specify if not
running within case_boot_lmer().}
}
\description{
Bootstrap resampling function that creates a bootstrapped sample of plots with replacement and
returns model output from trend_fun() for that sample. This is mostly an internal function run within case_boot_lmer(),
but can be used as a stand alone function.
}
\examples{
\dontrun{

fake_df <- data.frame(Plot_Name = rep(paste0(rep("PARK.", 9), 1:9), each = 3),
                      cycle = rep(1:3, times = 9),
                      resp = runif(27, 0, 20))

boot_ex <- case_boot_sample(fake_df, y = "resp", sample = TRUE)

}

}
