% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_slopes.R
\name{plot_slopes}
\alias{plot_slopes}
\title{plot_slopes: Plots slopes and 95\% CIs using output from case_boot_lmer}
\usage{
plot_slopes(
  df,
  ylabel,
  metric,
  order = NA,
  group = NA,
  sign_only = FALSE,
  legend_position = "none"
)
}
\arguments{
\item{df}{Data frame containing a column called Plot_Name, a column called cycle, and a column with at least one
response variable. If Unit_Code is in the data frame, then a facet by park will be plotted.}

\item{order}{Quoted column plot will sort by if specified.}

\item{group}{Group to color code by. If not specified, all points and lines will be black. If "Network" is specified,
then points and lines will be color coded by network, where ERMN is blue, MIDN is orange, NCRN is yellow, and NETN is green.}

\item{sign_only}{TRUE/FALSE. Denotes whether to plot all trends (FALSE; Default) or only significant trends (TRUE).}

\item{legend_position}{Quoted position for legend following ggplot positions. Default is 'none'.}

\item{xlab}{Quoted title for x axis.}

\item{ylab}{Quoted title for y axis.}
}
\description{
Function plots the slopes from case_boot_lmer output
}
