% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_response.R
\name{plot_trend_response}
\alias{plot_trend_response}
\title{plot_trend_response: Plots response with CIs using output from case_boot_lmer/loess}
\usage{
plot_trend_response(
  df,
  xlab,
  ylab,
  model_type = c("lmer", "loess"),
  group = NA,
  ribbon = FALSE,
  sign_color = c("#D3D3D3", "#696969", "#228B22", "#CD5C5C"),
  facet_scales = c("fixed"),
  facet_cols = 4,
  ptsize = 1
)
}
\arguments{
\item{df}{Data frame containing a column called Plot_Name, a column called cycle, and a column with at least one
response variable. If Unit_Code is in the data frame, then a facet by park will be plotted.}

\item{xlab}{Quoted title for x axis.}

\item{ylab}{Quoted title for y axis.}

\item{group}{Quoted column for facet wraps. If not specified, only 1 plot will be returned}

\item{ribbon}{Options are TRUE or FALSE (Default). If TRUE, will plot error as a ribbon instead of errorbars.}

\item{sign_color}{String of 4 colors to indicate a trend that is not modeled, not significant or significant increase
or significant decrease. Default is c("#D3D3D3", "#696969", "#228B22", "#CD5C5C"), which are light grey, dark grey,
forest green and a shade of red. Lines and outlines will be color coded based on this parameter. Ribbons and error bars
will only plot if trends were modeled (eg park had > 6 plots). Note that for the loess model, significance is determined by
comparing the confidence intervals of the first time step to the last time step. If there is no overlap, the trend is
considered significant.}

\item{facet_scales}{Options are "fixed" (Default), "free", "free_y", "free_x". Fixed means all axes will be
identical among facets. Free means axes will vary by facets.}

\item{facet_cols}{Number of columns for facet wrap. Default is 4.}

\item{ptsize}{Size of points to be plotted. Default is 1. Must be numeric.}
}
\description{
Function plots predicted responses for each time step in the data and 95% confidence intervals
around the response derived from case bootstrapping. For the lmer model, a trend is considered significant
if the confidence interval of the slope does not contain 0. For the loess model, a trend is considered significant
if the confidence intervals of the first and last time steps don't overlap. Note that specifying a window to test
for significance (e.g. compare first 4 years vs. last 4 years) is in development. Significant trends are plotted
as solid lines and filled symbols and a separate color (if specified) from non-significant trends. Non-significant
trends are plotted as dashed lines with open symbols. For models with insufficient sample size for case bootstrap
(<7 plots), the coefficients from the model on the raw data were plotted as dashed, light-grey lines without
confidence intervals.
}
\examples{
\dontrun{

#----- Dataset with 1 park -----
fake_df <- data.frame(Plot_Name = rep(paste0(rep("PARK.", 9), 1:9), each = 3),
                      cycle = rep(1:3, times = 9),
                      resp = runif(27, 0, 20))

boot1 <- case_boot_lmer(fake_df, y = "resp", num_reps = 10, random_type = 'intercept', chatty = TRUE)

plot_trend_response(boot1, xlab = 'cycle', ylab = 'BA', model_type = 'lmer') +
  scale_x_continuous(breaks = c(1, 2, 3), labels = c("1", "2", "3"))


#----- Dataset with 2 parks iterating through each park with purrr -----
# Create fake dataset
fake_2pk <- data.frame(Plot_Name = c(rep(paste0(rep("APRK-", 12), sprintf("\%02d", 1:12)), each = 3),
                       rep(paste0(rep("BPRK-", 12), sprintf("\%02d", 13:24)), each = 3)),
                       park = c(rep("APRK", 36), rep("BPRK", 36)),
                       cycle = rep(1:3, times = 24),
                       resp = runif(72, 0, 30))

# Nest dataset by park
nested_df <- fake_2pk \%>\% mutate(grp = park) \%>\% group_by(park) \%>\% nest()

# Run case_boot_lmer on nested dataset
boot2 <- nested_df \%>\% mutate(
  model = map(data, ~case_boot_loess(., x = "cycle", y = "resp", ID = "Plot_Name",
                                    span = 0.75, group = "grp",
                                    num_reps = 100, chatty = TRUE)))

# Compile results
boot_results <- boot2 \%>\% select(park, model) \%>\% unnest(model) \%>\% select(-num_boots)

# Plot results
plot_trend_response(boot_results, xlab = "Cycle", ylab = "BA", group = "park", ribbon = T,
                    facet_scales = "free") +
  scale_x_continuous(breaks = c(1, 2, 3), labels = c("1", "2", "3"))
}

}
