% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_response.R
\name{plot_trend_response}
\alias{plot_trend_response}
\title{plot_trend_response: Plots response with CIs using output from case_boot_lmer}
\usage{
plot_trend_response(df, ylab)
}
\arguments{
\item{df}{Data frame containing a column called Plot_Name, a column called cycle, and a column with at least one
response variable. If Unit_Code is in the data frame, then a facet by park will be plotted.}

\item{ylab}{Quoted title for y axis}
}
\description{
Function plots response of
}
\examples{
\dontrun{

fake_df <- data.frame(Plot_Name = rep(paste0(rep("PARK.", 9), 1:9), each = 3),
                      cycle = rep(1:3, times = 9),
                      resp = runif(27, 0, 20))

boot1 <- case_boot_lmer(fake_df, y = "resp", num_reps = 10, random_type = 'intercept', chatty = TRUE)

plot_trend_response(boot1, "Tree Basal Area")


}

}
