% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_fun.R
\name{trend_fun}
\alias{trend_fun}
\title{trend_fun: fit random intercept model and return results of interest}
\usage{
trend_fun(
  df,
  x = "cycle",
  y,
  ID = "Plot_Name",
  random_type = c("intercept", "slope")
)
}
\arguments{
\item{df}{Data frame containing a column called Plot_Name, a column called cycle, and a column with at least one
response variable.}

\item{x}{Quoted time variable for trend analysis. Default is "cycle", but can also model by year. Must be numeric.}

\item{y}{Quoted response variable in the data frame.}

\item{ID}{Quoted name of column containing site or plot IDs. Default is "Plot_Name", and assumes the first 4 characters
are a park code.}

\item{random_type}{intercept or slope. The intercept option (default) will fit a random intercept on plot with (1|Plot_Name) as
random component. The slope option will fit a random slope model with (1 + cycle|Plot_Name)}
}
\description{
fit random intercept model using Plot_Name, and return intercept, slope and predicted responses for
each time step (i.e. cycle or year) represented in the data. This is mostly an internal function run within case_boot_lmer(),
but can be used as a stand alone function.
}
\examples{
\dontrun{
library(magrittr)
fake_df <- data.frame(Plot_Name = rep(paste0(rep("PARK.", 9), 1:9), each = 3),
                      group = "PARK",
                      cycle = rep(1:3, times = 9),
                      resp = runif(27, 0, 20))

mod <- trend_fun(fake_df, x = "cycle", y = "resp", ID = "Plot_Name", random_type = "slope")
}

}
