% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_bootstrap_funs.R
\name{trend_fun}
\alias{trend_fun}
\title{trend_fun: fit random intercept model and return results of interest}
\usage{
trend_fun(df, y)
}
\arguments{
\item{df}{Data frame containing a column called Plot_Name, a column called cycle, and a column with at least one
response variable.}

\item{y}{Quoted response variable in the data frame.}
}
\description{
fit random intercept model using Plot_Name, and return intercept, slope and predicted responses for
each cycle represented in the data. This is mostly an internal function run within case_boot_lmer(),
but can be used as a stand alone function.
}
\examples{
\dontrun{
library(magrittr)
fake_df <- data.frame(Plot_Name = rep(paste0(rep("PARK.", 9), 1:9), each = 3),
                      cycle = rep(1:3, times = 9),
                      resp = runif(27, 0, 20))

mod <- trend_fun(fake_df, "resp")
}

}
